pyproject.toml
[tool.poetry]
name = "nofos"
version = "3.1.0"
description = "the no-code solo nofo web flow"
authors = ["Paul Craig <paul@pcraig.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.13"
django = "5.0.13"
beautifulsoup4 = "^4.12.2"
markdownify = "1.1.0"
django-easy-audit = "^1.3.5"
psycopg2 = "^2.9.10"
django-environ = "^0.12.0"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
python-slugify = "^8.0.1"
django-project-version = "^0.19.0"
docraptor = "^3.1.0"
django-constance = "^4.3.2"
django-picklefield = "^3.1"
markdown = "3.5.2"
requests = "^2.32.3"
cssutils = "^2.9.0"
martor = "^1.6.45"
django-mirror = "^0.1.13"
mammoth = "^1.9.0"
tomli = "^2.0.2"
django-ninja = "^1.3.0"
cryptography = "^44.0.2"
pyjwt = "^2.10.1"
google-cloud-secret-manager = "^2.23.2"
boto3 = "^1.38.12"

# sortable = "@3", check nofo_index.html

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
freezegun = "^1.3.1"
isort = "^6.0.1"

[tool.poetry.scripts]
start = 'bloom_nofos.scripts.server:main'
collectstatic = 'bloom_nofos.scripts.server:collectstatic'
migrate = 'bloom_nofos.scripts.server:migrate'
db-migrate = 'bloom_nofos.scripts.server:migrate'
makemigrations = 'bloom_nofos.scripts.server:makemigrations'
test = 'bloom_nofos.scripts.server:test'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
